Hotfix 'tensorflow' as interface kwarg in qnode
When checking whether the interface kwarg to a qnode is valid we allow
a few variations of the interface name, such as 'tf' and 'tensorflow'.
However, when dispatching to default.qubit subdevices, only one name
will properly access the subdevice, preventing the interface to
backpropagate through the device, instead defaulting to parameter-shift.